Get rid of loadUiType module and distribute the Qt imports
