--- scripts/SimplexUI/loadUiType.py
@@ -1,76 +1,0 @@
-'''
-Copyright 2016, Blur Studio
-
-This file is part of Simplex.
-
-Simplex is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Simplex is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with Simplex.  If not, see <http://www.gnu.org/licenses/>.
-'''
-
-import os
-import xml.etree.ElementTree as xml
-from cStringIO import StringIO
-
-def toPyObject(thing):
-	try:
-		return thing.toPyObject()
-	except:
-		return thing
-
-from Qt import QtGui, QtCore, QtWidgets, QtCompat
-from Qt.QtCore import Signal, QSortFilterProxyModel, Slot, QModelIndex
-from Qt.QtCore import Qt, QObject, QTimer, QPoint, QEvent, QItemSelection, QSettings
-from Qt.QtWidgets import QMessageBox, QInputDialog, QFileDialog, QMenu, QApplication, QAction
-from Qt.QtWidgets import QDialog, QMainWindow, QSplashScreen, QShortcut, QProgressDialog
-from Qt.QtGui import QStandardItemModel, QStandardItem, QKeySequence, QCursor, QMouseEvent
-try:
-	import pyside2uic as uic
-except ImportError:
-	uic = QtCompat.loadUi
-
-
-
-def loadUiType(fileVar, subFolder="ui", uiName=None):
-	"""
-	To use this, pass the path to a UI file, and capture the output
-	as FormClass and BaseClass.  Define your class inheriting from both
-	and call self.setupUi(self) in the __init__
-	"""
-	uiFolder, filename = os.path.split(fileVar)
-	if uiName is None:
-		uiName = os.path.splitext(filename)[0]
-	if subFolder:
-		uiFile = os.path.join(uiFolder, subFolder, uiName+".ui")
-
-	parsed = xml.parse(uiFile)
-	widget_class = parsed.find('widget').get('class')
-	form_class = parsed.find('class').text
-
-	with open(uiFile, 'r') as f:
-		o = StringIO()
-		frame = {}
-
-		uic.compileUi(f, o, indent=0)
-		pyc = compile(o.getvalue(), '<string>', 'exec')
-		exec pyc in frame
-
-		#Fetch the base_class and form class based on their type in the xml from designer
-		form_class = frame['Ui_{0}'.format(form_class)]
-
-		try:
-			base_class = eval('QtGui.{0}'.format(widget_class))
-		except:
-			base_class = eval('QtWidgets.{0}'.format(widget_class))
-
-	return form_class, base_class
-
