+++ scripts/SimplexUI/simplexdialog.py
@@ -118,17 +118,17 @@ from fnmatch import fnmatchcase
 from functools import wraps
 
 try:
-	# These modules are unique to Blur Studios 
+	# These modules are unique to Blur Studios
 	import blurdev
 except ImportError:
 	blurdev = None
 
-# This module imports QT from PyQt4, PySide or PySide2
-# Depending on what's available
-from loadUiType import (loadUiType, toPyObject, QMessageBox, QMenu, QApplication,
-						Slot, QSortFilterProxyModel, QMainWindow, QInputDialog, QSettings,
-						QFileDialog, QShortcut, Qt, QObject, QTimer, QItemSelection,
-						QStandardItemModel, QStandardItem, QModelIndex, QKeySequence, QProgressDialog)
+from Qt import QtGui, QtCore, QtWidgets, QtCompat
+from Qt.QtCore import Signal, QSortFilterProxyModel, Slot, QModelIndex
+from Qt.QtCore import Qt, QObject, QTimer, QPoint, QEvent, QItemSelection, QSettings
+from Qt.QtWidgets import QMessageBox, QInputDialog, QFileDialog, QMenu, QApplication, QAction
+from Qt.QtWidgets import QDialog, QMainWindow, QSplashScreen, QShortcut, QProgressDialog
+from Qt.QtGui import QStandardItemModel, QStandardItem, QKeySequence, QCursor, QMouseEvent
 
 from dragFilter import DragFilter
 
@@ -139,7 +139,10 @@ from constants import (PRECISION, COLUMNCOUNT, THING_ROLE, VALUE_ROLE, WEIGHT_RO
 					   TYPE_ROLE, PARENT_ROLE, THING_NAME_COL, SLIDER_VALUE_COL,
 					   SHAPE_WEIGHT_COL, S_SHAPE_TYPE, S_SLIDER_TYPE, S_GROUP_TYPE,
 					   S_SYSTEM_TYPE, C_SHAPE_TYPE, C_SHAPE_PAR_TYPE, C_SLIDER_TYPE,
-					   C_SLIDER_PAR_TYPE, C_COMBO_TYPE, C_GROUP_TYPE, C_SYSTEM_TYPE,) 
+					   C_SLIDER_PAR_TYPE, C_COMBO_TYPE, C_GROUP_TYPE, C_SYSTEM_TYPE,)
+
+import utils
+from utils import toPyObject
 
 # If the decorated method is a slot for some Qt Signal
 # and the method signature is *NOT* the same as the
@@ -213,16 +216,11 @@ class singleShot(QObject):
 		self._function(inst, args)
 
 
-
-
-
-
-# LOAD THE UI base classes
-FormClass, BaseClass = loadUiType(__file__)
-class SimplexDialog(FormClass, BaseClass):
+class SimplexDialog(QMainWindow):
 	def __init__(self, parent=None, dispatch=None):
 		super(SimplexDialog, self).__init__(parent)
-		self.setupUi(self)
+		QtCompat.loadUi(utils.getUiFile(__file__), self)
+		# self.setupUi(self)
 
 		self._sliderMenu = None
 		self._comboMenu = None
@@ -882,7 +880,7 @@ class SimplexDialog(FormClass, BaseClass):
 	def loadObject(self, obj):
 		if not obj:
 			return
-			
+
 		self.uiCurrentSystemCBOX.clear()
 		objName = System.getObjectName(obj)
 		self._currentObject = obj
@@ -979,7 +977,7 @@ class SimplexDialog(FormClass, BaseClass):
 	def clearCurrentSystem(self):
 		self.system = System()
 		self.toolActions.system = self.system
-		self.forceSimplexUpdate()	
+		self.forceSimplexUpdate()
 
 	# Falloff Editing
 	def buildFalloffLists(self):
@@ -1194,7 +1192,7 @@ class SimplexDialog(FormClass, BaseClass):
 		thing = toPyObject(selItems[0].data(THING_ROLE))
 		oldName = thing.name
 		if oldName == userName: #Don't double-change the name
-			return 
+			return
 		nn = self.getNextName(userName, sliderNames)
 		sliderNames.append(nn)
 		self.renameSlider(thing, nn)
@@ -1442,7 +1440,7 @@ class SimplexDialog(FormClass, BaseClass):
 		for c in combos:
 			comboItems.extend(self._comboTreeMap[c])
 
-		# check if sliders are part of combos, asks the user whether 
+		# check if sliders are part of combos, asks the user whether
 		# they want to delete those combos, and returns the list
 		if comboItems:
 			comboThings = [toPyObject(i.data(THING_ROLE))for i in comboItems]
@@ -2011,7 +2009,7 @@ class SimplexDialog(FormClass, BaseClass):
 			thing.expanded = tree.isExpanded(index)
 			if isinstance(thing, Simplex):
 				if tree == self.uiComboTREE:
-					thing.comboExpanded = tree.isExpanded(index) 
+					thing.comboExpanded = tree.isExpanded(index)
 				elif tree == self.uiSliderTREE:
 					thing.sliderExpanded = tree.isExpanded(index)
