+++ scripts/SimplexUI/utils.py
@@ -1,0 +1,18 @@
+"""Utility functions."""
+import os
+
+def toPyObject(thing):
+  try:
+	  return thing.toPyObject()
+  except:
+	  return thing
+
+
+def getUiFile(fileVar, subFolder="ui", uiName=None):
+	"""Get the path to the .ui file"""
+	uiFolder, filename = os.path.split(fileVar)
+	if uiName is None:
+		uiName = os.path.splitext(filename)[0]
+	if subFolder:
+		uiFile = os.path.join(uiFolder, subFolder, uiName+".ui")
+	return uiFile
